+++ src/Behaviour-Trees/Assets/Plugins/BehaviourTrees/Runtime/Behaviours/SequenceBehaviour.cs
@@ -1,0 +1,8 @@
+﻿using Derrixx.BehaviourTrees.Runtime.Pure.Composites;
+
+namespace Derrixx.BehaviourTrees.Runtime.Behaviours
+{
+	public sealed class SequenceBehaviour : CompositeBehaviour<Sequence>
+	{
+	}
+}