+++ src/Behaviour-Trees/Assets/Plugins/BehaviourTrees/Runtime/INode.cs
@@ -1,0 +1,13 @@
+﻿using Derrixx.BehaviourTrees.Runtime.Pure;
+
+namespace Derrixx.BehaviourTrees.Runtime
+{
+	public interface INode
+	{
+		INode Parent { get; set; }
+		
+		NodeState Execute();
+
+		void InjectBlackboard(IBlackboard blackboard);
+	}
+}