+++ src/Behaviour-Trees/Assets/Plugins/BehaviourTrees/Runtime/Pure/Decorators/ExecuteNTimes.cs
@@ -1,4 +1,4 @@
-﻿namespace Derrixx.BehaviourTrees.Runtime.Decorators
+﻿namespace Derrixx.BehaviourTrees.Runtime.Pure.Decorators
 {
     public sealed class ExecuteNTimes : Decorator
     {
@@ -6,28 +6,28 @@
         
         public int Counter { get; private set; }
         
-        public ExecuteNTimes(int n, Node child) : base($"Execute {n} times", child)
+        public ExecuteNTimes(int n, INode child) : base($"Execute {n} times", child)
         {
             _limit = n;
         }
 
-        public override NodeState Execute(IBlackboard blackboard)
+        public override NodeState Execute()
         {
             if (Counter >= _limit)
                 return NodeState.Success;
                 
             Counter++;
-            return Child.Execute(blackboard);
+            return Child.Execute();
         }
 
         public void Reset() => Counter = 0;
 
-        protected internal override string PrintNode(int nodeLevel)
+        /*protected internal override string PrintNode(int nodeLevel)
         {
             string output = base.PrintNode(nodeLevel);
             output += Child.PrintNode(nodeLevel + 1);
             
             return output;
-        }
+        }*/
     }
 }